feat: remove quest button and button unused components
