+++ apps/govquests-frontend/src/components/ui/Button.tsx
@@ -1,61 +1,57 @@
-import { cn } from "@/lib/utils";
-import type { LucideIcon } from "lucide-react";
-import type React from "react";
-import LoadingIndicator from "./LoadingIndicator";
-import Spinner from "./Spinner";
+import * as React from "react";
+import { Slot } from "@radix-ui/react-slot";
+import { cva, type VariantProps } from "class-variance-authority";
 
-interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  variant?: "primary" | "secondary";
-  size?: "sm" | "md" | "lg";
-  icon?: LucideIcon;
-  loading?: boolean;
-  className?: string;
+import { cn } from "@/lib/utils";
+
+const buttonVariants = cva(
+  "inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0",
+  {
+    variants: {
+      variant: {
+        default:
+          "bg-primary text-primary-foreground transition-all duration-200 hover:invert border-2 border-primary-foreground shadow-[2px_3px_0_0_rgba(0,0,0)]",
+        destructive:
+          "bg-destructive text-destructive-foreground shadow-sm hover:bg-destructive/90",
+        outline:
+          "border border-input bg-background shadow-sm hover:bg-accent hover:text-accent-foreground",
+        secondary:
+          "bg-background/90 text-primary border border-primary/10 hover:border-primary/30 hover:border-2 box-border",
+        ghost: "hover:bg-accent hover:text-accent-foreground",
+        link: "text-primary underline-offset-4 hover:underline",
+      },
+      size: {
+        default: "h-9 px-4 py-2",
+        sm: "h-8 rounded-md px-3 text-xs",
+        lg: "h-10 rounded-md px-8",
+        icon: "h-9 w-9",
+      },
+    },
+    defaultVariants: {
+      variant: "default",
+      size: "default",
+    },
+  },
+);
+
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
+    VariantProps<typeof buttonVariants> {
+  asChild?: boolean;
 }
 
-const Button: React.FC<ButtonProps> = ({
-  children,
-  variant = "primary",
-  size = "md",
-  icon: Icon,
-  loading,
-  className,
-  ...props
-}) => {
-  const baseClasses =
-    "flex items-center justify-center font-medium rounded-md transition-colors";
+const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
+  ({ className, variant, size, asChild = false, ...props }, ref) => {
+    const Comp = asChild ? Slot : "button";
+    return (
+      <Comp
+        className={cn(buttonVariants({ variant, size, className }))}
+        ref={ref}
+        {...props}
+      />
+    );
+  },
+);
+Button.displayName = "Button";
 
-  const variantClasses = {
-    primary:
-      "bg-primary text-primary-foreground hover:bg-primary/70 transition",
-    secondary:
-      "bg-secondary text-secondary-foreground hover:bg-secondary/70 transition",
-  };
-  const sizeClasses = {
-    sm: "px-3 py-1 text-sm",
-    md: "px-4 py-2",
-    lg: "px-6 py-3 text-lg",
-  };
-
-  return (
-    <button
-      className={cn(
-        baseClasses,
-        variantClasses[variant],
-        sizeClasses[size],
-        className,
-      )}
-      {...props}
-    >
-      {loading ? (
-        <Spinner />
-      ) : (
-        <>
-          {children}
-          {Icon && <Icon className={cn("w-5 h-5", children && "ml-2")} />}
-        </>
-      )}
-    </button>
-  );
-};
-
-export default Button;
+export { Button, buttonVariants };
