+++ apps/govquests-frontend/src/domains/authentication/components/SignInButton.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Button } from "@/components/ui/shadcn-button";
+import { Button } from "@/components/ui/button";
 import Spinner from "@/components/ui/Spinner";
 import { ConnectKitButton, useSIWE } from "connectkit";
 import { ChevronDown } from "lucide-react";
